Create Longest common substring value of two numbers
Given two integers N and M. The task is to find the longest contiguous subset in binary representation of both the numbers and print its decimal value. If there are two or more longest common substrings with the same length, the print the maximum value of all the common substrings.